patch from gpl ghostscript 8.50 (12/16) : src

+++ src/spdiffx.h
@@ -17,7 +17,7 @@
   
 */
 
-/* $Id: spdiffx.h,v 1.4 2005/12/13 16:57:28 jemarch Exp $ */
+/* $Id: spdiffx.h,v 1.5 2006/03/08 12:30:25 Arabidopsis Exp $ */
 /* Definitions for PixelDifference filters */
 /* Requires strimpl.h */
 
@@ -43,7 +43,7 @@ typedef struct stream_PDiff_state_s {
     int case_index;		/* switch index for case dispatch */
     /* The following are updated dynamically. */
     uint row_left;		/* # of bytes left in row */
-    byte prev[s_PDiff_max_Colors];	/* previous sample */
+    uint prev[s_PDiff_max_Colors];	/* previous sample */
 } stream_PDiff_state;
 
 #define private_st_PDiff_state()	/* in spdiff.c */\
